+++ client/pages/login/index.tsx
@@ -5,7 +5,7 @@ import Image from 'next/image';
 import Link from 'next/link';
 
 import sideImage from '@assets/images/login-side-test.png';
-import logo from '@assets/images/logo.png';
+import logo from '@assets/images/logo.webp';
 import Login from '@components/Login';
 
 const LoginPage = () => {
